requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dacite==1.9.2
Deprecated==1.2.18
fastapi==0.103.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jaconv==0.4.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
mutagen==1.47.0
pathlib==1.0.1
platformdirs==4.3.6
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pykakasi==2.3.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-telegram-bot==21.10
pytube==15.0.0
RapidFuzz==3.12.1
redis==5.2.1
requests==2.32.3
rich==13.9.4
six==1.17.0
sniffio==1.3.1
soundcloud-v2==1.6.0
soupsieve==2.6
spotdl==4.2.11
spotipy==2.25.0
SQLAlchemy==2.0.38
starlette==0.27.0
syncedlyrics==1.0.1
text-unidecode==1.3
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.3.0
uvicorn==0.23.2
websockets==14.2
wrapt==1.17.2
yandex-music==2.2.0
yarl==1.18.3
yt-dlp==2025.2.19
ytmusicapi==1.10.1
